First commit to mac branch. Note this branch is temporary and will be merged into master once it's tested to not break the windows version.
- First commit! It compiles in Xcode. No interface yet.
- All #include stdafx.h calls are now wrapped in #ifdef _WIN32.
- Lots of little changes to make the code compliant with C++11 in clang++. Hopefully didn't break the Visual Studio build too much.
- added src code for tinyxml and minizip (minizip isn't new, it's just already compiled into zlibstat.lib... that's confusing and not standard. let's fix that).
- Added a new osdepend function LogDebug to replace all the calls to OutputDebugString since it's a windows function.

+++ src/formats/SquarePS2.h
@@ -1,11 +1,10 @@
 #pragma once
 #include "VGMSeq.h"
 
-class BGMSeq :
-	public VGMSeq
+class BGMSeq : public VGMSeq
 {
 public:
-	BGMSeq(RawFile* file, ULONG offset);
+	BGMSeq(RawFile* file, uint32_t offset);
 public:
 	virtual ~BGMSeq(void);
 
@@ -18,8 +17,7 @@ protected:
 };
 
 
-class BGMTrack
-	: public SeqTrack
+class BGMTrack : public SeqTrack
 {
 public:
 	BGMTrack(BGMSeq* parentSeq, long offset = 0, long length = 0);
